[Refactor] auth api, validate 로직 분리, 연동
